% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_derivatives.R
\name{weather_derivative_functions}
\alias{weather_derivative_functions}
\alias{derive_dew_point}
\alias{derive_vpd}
\alias{derive_temperature_range}
\alias{derive_leaf_wetness_from_rh}
\alias{derive_leaf_wetness_from_rh_temp}
\alias{derive_favorable_condition}
\title{Derive meteorological variables}
\usage{
derive_dew_point(data, temp, rh, name = "dew_point")

derive_vpd(data, temp, rh, name = "vpd")

derive_temperature_range(data, tmax, tmin, name = "temp_range")

derive_leaf_wetness_from_rh(
  data,
  rh,
  threshold = 90,
  name = "leaf_wetness_est"
)

derive_leaf_wetness_from_rh_temp(
  data,
  rh,
  temp,
  rh_threshold = 90,
  temp_range = NULL,
  name = "leaf_wetness_est"
)

derive_favorable_condition(data, condition, name = "favorable")
}
\arguments{
\item{data}{A data frame.}

\item{temp, rh, tmax, tmin}{Columns in \code{data}.}

\item{name}{Name of the new column.}

\item{threshold, rh_threshold}{Numeric threshold used to create binary wetness
or favorability variables.}

\item{temp_range}{Optional numeric vector of length two. When supplied,
estimated leaf wetness is one only when RH is high and temperature is inside
this range.}

\item{condition}{A logical expression evaluated in \code{data}.}
}
\value{
A data frame with the new derived column.
}
\description{
These functions add biologically useful variables to a weather data frame.
Column arguments are supplied unquoted. The original data frame is returned
with one additional column.
}
\details{
\code{derive_dew_point()} uses the Magnus approximation with constants
\code{a = 17.625} and \code{b = 243.04}. \code{derive_vpd()} computes saturation vapor
pressure from temperature and actual vapor pressure from relative humidity.
Relative humidity is interpreted as percent from 0 to 100, and VPD is returned
in kPa. Dew point is returned as \code{NA_real_} when relative humidity is zero
because the logarithmic Magnus approximation has no finite value at zero.
}
\examples{
weather <- data.frame(temp = c(20, 24), rh = c(85, 95), rain = c(0, 1))

weather |>
  derive_dew_point(temp, rh) |>
  derive_vpd(temp, rh) |>
  derive_leaf_wetness_from_rh(rh, threshold = 90)

}
