% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windcut-package.R
\docType{package}
\name{windcut-package}
\alias{windcut}
\alias{windcut-package}
\title{windcut: Weather Time Series Feature Engineering for Plant Disease Epidemiology}
\description{
\code{windcut} provides functions to validate weather series, generate candidate
relative-time windows, summarise weather inside those windows, simulate
weather and assessment datasets, and create model-ready feature tables for
plant disease epidemiology workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlvesKS/windcut}
  \item \url{https://alvesks.github.io/windcut/}
  \item Report bugs at \url{https://github.com/AlvesKS/windcut/issues}
}

}
\author{
\strong{Maintainer}: Kaique S. Alves \email{kaique.alves@example.com}

}
\keyword{internal}
